refactor: make all koda-cli modules pub(crate)
The 29 modules were pub solely to support the now-abandoned approach of
using docs.rs/koda-cli as a user manual. With issue #728 tracking the
move to mdBook, there is no reason for these internals to be public.

Approach: expose one pub fn run() from the library; make main.rs a
5-liner that calls it. All CLI argument parsing + dispatch moves into
pub(crate) mod app. Every other module stays pub(crate).

Changes:
- koda-cli/src/main.rs         → 9-line shell: calls koda_cli::run()
- koda-cli/src/app.rs          → new: contains former main() logic
- koda-cli/src/lib.rs          → pub mod → pub(crate) mod (all 29)
                                  + pub async fn run() entry point
                                  + pub(crate) mod app
- koda-cli/src/repl.rs         → repl_commands tests inlined here
                                  (moved from regression_test.rs)
- koda-cli/tests/regression_test.rs → repl_commands module removed,
                                       doc comment updated
- koda-cli/src/tool_history.rs → dead is_empty() deleted (YAGNI:
                                  only existed because the method
                                  was pub; no caller exists)
