+++ koda-cli/src/tool_history.rs
@@ -53,10 +53,6 @@ impl ToolOutputHistory {
     pub fn len(&self) -> usize {
         self.entries.len()
     }
-
-    pub fn is_empty(&self) -> bool {
-        self.entries.is_empty()
-    }
 }
 
 #[cfg(test)]
